Move validator key and consts
